chore: remove unused component and deps
