+++ package.json
@@ -68,7 +68,6 @@
     "@radix-ui/react-icons": "^1.1.1",
     "@radix-ui/react-id": "^1.0.0",
     "@radix-ui/react-label": "^2.0.0",
-    "@radix-ui/react-menu": "^2.0.1",
     "@radix-ui/react-navigation-menu": "^1.2.0",
     "@radix-ui/react-popover": "^1.0.2",
     "@radix-ui/react-portal": "^1.0.1",
